Add the latest version, some features are not complete yet

+++ istream-backend/src/database/defaultModules/Server/Nginx Dash/run.sh
@@ -22,6 +22,8 @@ docker ps -a -q --filter "name=istream_server_nginx_container" | grep -q . &&
 
 docker run --name istream_server_nginx_container ${dockerCupConfig} ${dockerMemoryConfig} -p ${serverDockerPort}:80 -d istream_server_nginx_image
 
-docker cp "${DIR}/Run/Videos/." istream_server_nginx_container:/usr/local/nginx/html/
-docker cp "${DIR}/Run/config.sh" istream_server_nginx_container:/usr/local/nginx/html/
-docker exec istream_server_nginx_container bash /usr/local/nginx/html/config.sh
+sh "${DIR}/Run/fetcher.sh"
+if [ -f "${DIR}/Run/config.sh" ]; then
+    docker cp "${DIR}/Run/config.sh" istream_server_nginx_container:/usr/local/nginx/html/
+    docker exec istream_server_nginx_container bash /usr/local/nginx/html/config.sh
+fi
